applywise/apps/ai-service/trial/requirements.txt
flask
flask-cors
joblib
pdfplumber
reportlab
google-generativeai
python-dotenv
nltk
pandas
youtube-search-python
requests
beautifulsoup4
selenium
webdriver-manager
